Restructure into src layout and add PyPI packaging metadata
Moves source into src/build123d_mcp/ package so the wheel builds
correctly. Updates all internal imports, test imports, console script
entry point, and .mcp.json. Adds hatchling build backend, MIT licence,
author, keywords, classifiers, and project URLs to pyproject.toml.
uv build now produces a clean wheel ready for uv publish.
