+++ src/build123d_mcp/tools/interference.py
