+++ src/build123d_mcp/tools/measure.py
