+++ src/build123d_mcp/tools/render.py
